+++ push_swap.h
@@ -6,8 +6,8 @@
 
 typedef struct s_stack
 {
-	long			num;
-	long			idx;
+	int			num;
+	int			idx;
 	struct s_stack	*next;
 	struct s_stack	*prev;
 }	t_stack;
@@ -17,15 +17,15 @@ typedef struct s_stack
 # include <stdio.h>
 
 void	error();
-long    assign_idx(char **results, long num);
-t_stack	*new_node(long num, char **results);
+int    assign_idx(char **results, int num);
+t_stack	*new_node(int num, char **results);
 void 	append_node(t_stack **head, t_stack *node);
 t_stack	**get_stack(char **results);
 int		is_duplicated(t_stack **a);
 int     is_valid_input(t_stack **a);
 t_stack	*lst_last(t_stack *lst);
 int		is_sorted(t_stack **a);
-long	stack_size(t_stack **a);
+int	stack_size(t_stack **a);
 
 void	sa(t_stack **a);
 void	sb(t_stack **b);
@@ -39,11 +39,11 @@ void    rra(t_stack **a);
 void    rrb(t_stack **b);
 void    rrr(t_stack **a, t_stack **b);
 
-long    min_num(t_stack **stack);
+int    min_num(t_stack **stack);
 int	    find_pos(t_stack **a, int value);
 void    sort_three(t_stack **a);
 void    small_sort(t_stack **a);
-long    get_max_pos(t_stack *stack);
+int    get_max_pos(t_stack *stack);
 void	butterfly_sort(t_stack **a, t_stack **b);
 
 #endif
